server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
azure-cognitiveservices-language-textanalytics==5.3.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
aiofiles==23.2.1